accessible/tests/mochitest/test_invalidate_elmli.html

<!DOCTYPE html>
<html>

<head>
  <title>Test HTML li and listitem bullet accessible cache</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/common.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/role.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/events.js"></script>

  <script type="application/javascript">

    ////////////////////////////////////////////////////////////////////////////
    // Helpers

    function testDefunctAccessible(aAcc, aNodeOrId)
    {
      if (aNodeOrId)
        ok(!isAccessible(aNodeOrId),
           "Accessible for " + aNodeOrId + " wasn't properly shut down!");

      ok(!aAcc.firstChild, "There is first child for shut down accessible!");
      ok(!aAcc.lastChild, "There is last child for shut down accessible!");
      ok(!aAcc.children.length, "There are children for shut down accessible!");

      // nextSibling
      var success = false;
      try {
        aAcc.nextSibling;
      } catch (e) {
        success = (e.result == Components.results.NS_ERROR_FAILURE);
      }

      ok(success, "There is next sibling for shut down accessible!");

      // previousSibling
      var success = false;
      try {
        aAcc.previousSibling;
      } catch (e) {
        success = (e.result == Components.results.NS_ERROR_FAILURE);
      }

      ok(success, "There is previous sibling for shut down accessible!");

      // parent
      var success = false;
      try {
        aAcc.parent;
      } catch (e) {
        success = (e.result == Components.results.NS_ERROR_FAILURE);
      }

      ok(success, "There is parent for shut down accessible!");
    }

    function testLiAccessibleTree()
    {
      // Test accessible tree.
      var accTree = {
        role: ROLE_LISTITEM,
        children: [
          {
            role: ROLE_STATICTEXT,
            children: []
          },
          {
            role: ROLE_TEXT_LEAF,
            children: []
          }
        ]
      };

      testAccessibleTree("li", accTree);
    }

    ////////////////////////////////////////////////////////////////////////////
    // Sequence item processors

    function hideProcessor()
    {
      this.liNode = getNode("li");
      this.li = getAccessible(this.liNode);
      this.bullet = this.li.firstChild;

      this.process = function hideProcessor_process()
      {
        this.liNode.style.display = "none";
      }

      this.onProcessed = function hideProcessor_onProcessed()
      {
        window.setTimeout(
          function(aArg1, aArg2)
          {
            testDefunctAccessible(aArg1, aArg2);
            gSequence.processNext();
          },
          0, this.li, this.liNode
        );
      }
    };

    function showProcessor()
    {
      this.liNode = getNode("li");

      this.process = function showProcessor_process()
      {
        this.liNode.style.display = "list-item";
      }

      this.onProcessed = function showProcessor_onProcessed()
      {
        testLiAccessibleTree();
        SimpleTest.finish();
      }
    };

    ////////////////////////////////////////////////////////////////////////////
    // Test

    var gSequence = null;

    function doTest()
    {
      testLiAccessibleTree();

      gSequence = new sequence();

      gSequence.append(new hideProcessor(), EVENT_HIDE, getAccessible("li"),
                       "hide HTML li");
      gSequence.append(new showProcessor(), EVENT_SHOW, getNode("li"),
                       "show HTML li");

      gSequence.processNext(); // SimpleTest.finish() will be called in the end
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="setParent shouldn't be virtual"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=496783">Mozilla Bug 496783</a>

  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <ul>
    <li id="li">item1</li>
  </ul>
</body>
</html>
